feat(linux): add multiple kernel patches
This commit add multiple patches to the Linux kernel, in order to
make it more appropriate to desktop use.

We start adding a patch to enable `-march=native` support to both
Intel and AMD CPUs, which only makes sense when building the kernel
in the same host that it will be executed.

We also add multiple patches from the Intel's Clear Linux project.

While we also added zen-kernel patches to the tree, they aren't used
yet, as we need to properly update them to the kernel version we use.

See each patch file to a more appropriate discussion of what each does.

Signed-off-by: Luis Holanda <luiscmholanda@gmail.com>