feat: start laptop support
This commit adds the base for laptop support, starting with
power configuration.

The base infrastructure is the `host.hardware.isLaptop` option,
that says if the host is a laptop. This option will be used in
other modules to provide laptop-specific configuration.

We then start these specific modules with a power management
module, which enables both upower and powertop while also
configuring systemd behavior for hibernation and lid switch.

Signed-off-by: Luis Holanda <luiscmholanda@gmail.com>