feat(kernel): use zen-kernel as base kernel
Removes the usage of the hardened kernel as base and use
zen-kernel. This simplifies patch management, as we don't
need to rework each patch to make it work with our current
version of Linux.

In the future, we can apply the patches from the hardened
version if we see something that makes sense.

Signed-off-by: Luis Holanda <luiscmholanda@gmail.com>