fix(time): workaround localtimed issues
For some unknown reason ([1] and [2]), `localtimed` is not
setting the time zone correctly if no previous set value
exists.

Workaround this issue by forcing a base value, which may or
may not be updated later by the service. There isn't much
we can do other than wait for the issue to be resolved.

[1]: https://github.com/NixOS/nixpkgs/issues/67673
[2]: https://github.com/NixOS/nixpkgs/issues/68489

Signed-off-by: Luis Holanda <luiscmholanda@gmail.com>